refactor(client): decompose modal from table
